+++ lib/src/base_cubit.dart
@@ -2,7 +2,7 @@ import 'dart:developer';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:spargo_tools/src/app_base_settings.dart';
-import 'package:spargo_tools/src/http_client.dart';
+import 'package:spargo_tools/src/http_client/http_client.dart';
 
 import 'exceptions/exceptions.dart';
 
@@ -31,17 +31,20 @@ abstract class BaseCubit<BaseState> extends Cubit<BaseState> {
     Function(ApiResponseData<ResponseType>)? callbackWhenError;
     Function(ApiResponseData<ResponseType>)? callbackWhenSuccess;
     if (onSuccess != null) {
-      assert(onError != null, 'Если есть callback onSuccess, то тогда нужно передавать и onError');
+      assert(onError != null,
+          'Если есть callback onSuccess, то тогда нужно передавать и onError');
 
       if (loadingState != null) {
         emit(loadingState);
       }
 
       callbackWhenError = (response) => onError!.call(response.exception);
-      callbackWhenSuccess = (response) => onSuccess.call(response.response as ResponseType);
+      callbackWhenSuccess =
+          (response) => onSuccess.call(response.response as ResponseType);
     } else {
       if (loadedState != null) {
-        assert(errorState != null, 'Если есть loadedState, то тогда нужно передавать и errorState');
+        assert(errorState != null,
+            'Если есть loadedState, то тогда нужно передавать и errorState');
       }
 
       if (loadingState != null) {
@@ -57,7 +60,8 @@ abstract class BaseCubit<BaseState> extends Cubit<BaseState> {
         };
       }
       if (loadedState != null) {
-        callbackWhenSuccess = (response) => emit(loadedState.call(response.response as ResponseType));
+        callbackWhenSuccess = (response) =>
+            emit(loadedState.call(response.response as ResponseType));
       }
     }
 
@@ -77,7 +81,8 @@ abstract class BaseCubit<BaseState> extends Cubit<BaseState> {
     return null;
   }
 
-  @Deprecated('Устарело, используйте apiRequest, чтобы получать детальную информацию об ошибках')
+  @Deprecated(
+      'Устарело, используйте apiRequest, чтобы получать детальную информацию об ошибках')
   Future<ResponseType?> baseRequest<ResponseType>(
     Future<ResponseType> request, {
     Function(String message, ExceptionType type)? onError,
@@ -89,17 +94,21 @@ abstract class BaseCubit<BaseState> extends Cubit<BaseState> {
     Function(ApiRequestData<ResponseType>)? callbackWhenError;
     Function(ApiRequestData<ResponseType>)? callbackWhenSuccess;
     if (onSuccess != null) {
-      assert(onError != null, 'Если есть callback onSuccess, то тогда нужно передавать и onError');
+      assert(onError != null,
+          'Если есть callback onSuccess, то тогда нужно передавать и onError');
 
       if (loadingState != null) {
         emit(loadingState);
       }
 
-      callbackWhenError = (response) => onError!.call(response.errorMessage!, response.errorType!);
-      callbackWhenSuccess = (response) => onSuccess.call(response.response as ResponseType);
+      callbackWhenError = (response) =>
+          onError!.call(response.errorMessage!, response.errorType!);
+      callbackWhenSuccess =
+          (response) => onSuccess.call(response.response as ResponseType);
     } else {
       if (loadedState != null) {
-        assert(errorState != null, 'Если есть loadedState, то тогда нужно передавать и errorState');
+        assert(errorState != null,
+            'Если есть loadedState, то тогда нужно передавать и errorState');
       }
 
       if (loadingState != null) {
@@ -115,7 +124,8 @@ abstract class BaseCubit<BaseState> extends Cubit<BaseState> {
         };
       }
       if (loadedState != null) {
-        callbackWhenSuccess = (response) => emit(loadedState.call(response.response as ResponseType));
+        callbackWhenSuccess = (response) =>
+            emit(loadedState.call(response.response as ResponseType));
       }
     }
 
@@ -147,24 +157,31 @@ abstract class BaseCubit<BaseState> extends Cubit<BaseState> {
     Function(ApiRequestData<ResponseType>) callbackWhenError;
     Function(ApiRequestData<ResponseType>) callbackWhenSuccess;
     if (onSuccess != null) {
-      assert(onError != null, 'Если есть callback onSuccess, то тогда нужно передавать и onError');
+      assert(onError != null,
+          'Если есть callback onSuccess, то тогда нужно передавать и onError');
 
       if (loadingState != null) {
         emit(loadingState);
       }
 
-      callbackWhenError = (response) => onError!.call(response.errorMessage!, response.errorType!);
-      callbackWhenSuccess = (response) => onSuccess.call(response.response as ResponseType);
+      callbackWhenError = (response) =>
+          onError!.call(response.errorMessage!, response.errorType!);
+      callbackWhenSuccess =
+          (response) => onSuccess.call(response.response as ResponseType);
     } else {
-      assert(loadedState != null, "Если не передавать callBack'и, то нужно передать все state'ы");
-      assert(errorState != null, "Если не передавать callBack'и, то нужно передать все state'ы");
+      assert(loadedState != null,
+          "Если не передавать callBack'и, то нужно передать все state'ы");
+      assert(errorState != null,
+          "Если не передавать callBack'и, то нужно передать все state'ы");
       //assert(loadingState != null, "Если не передавать callBack'и, то нужно передать все state'ы");
 
       if (loadingState != null) {
         emit(loadingState);
       }
-      callbackWhenError = (response) => emit(errorState!.call(response.errorMessage!, response.errorType!));
-      callbackWhenSuccess = (response) => emit(loadedState!.call(response.response));
+      callbackWhenError = (response) =>
+          emit(errorState!.call(response.errorMessage!, response.errorType!));
+      callbackWhenSuccess =
+          (response) => emit(loadedState!.call(response.response));
     }
 
     final response = await ApiRequest.requestOld(request);
@@ -187,20 +204,25 @@ abstract class BaseCubit<BaseState> extends Cubit<BaseState> {
     BaseState Function(List<ResponseType>? response)? loadedState,
     BaseState Function(String message, ExceptionType type)? errorState,
   }) async {
-    Function(ApiMultiRequestData<ResponseType>) callbackWhenError = (response) => {};
-    Function(ApiMultiRequestData<ResponseType>) callbackWhenSuccess = (response) => {};
+    Function(ApiMultiRequestData<ResponseType>) callbackWhenError =
+        (response) => {};
+    Function(ApiMultiRequestData<ResponseType>) callbackWhenSuccess =
+        (response) => {};
     if (onSuccess != null) {
-      assert(onError != null, 'Если есть callback onSuccess, то тогда нужно передавать и onError');
+      assert(onError != null,
+          'Если есть callback onSuccess, то тогда нужно передавать и onError');
 
       if (loadingState != null) {
         emit(loadingState);
       }
 
-      callbackWhenError = (response) => onError!.call(response.errorMessage!, response.errorType!);
+      callbackWhenError = (response) =>
+          onError!.call(response.errorMessage!, response.errorType!);
       callbackWhenSuccess = (response) => onSuccess.call(response.response!);
     } else {
       if (loadedState != null) {
-        assert(errorState != null, 'Если есть loadedState, то тогда нужно передавать и errorState');
+        assert(errorState != null,
+            'Если есть loadedState, то тогда нужно передавать и errorState');
       }
 
       if (loadingState != null) {
@@ -208,10 +230,12 @@ abstract class BaseCubit<BaseState> extends Cubit<BaseState> {
       }
 
       if (errorState != null) {
-        callbackWhenError = (response) => emit(errorState.call(response.errorMessage!, response.errorType!));
+        callbackWhenError = (response) =>
+            emit(errorState.call(response.errorMessage!, response.errorType!));
       }
       if (loadedState != null) {
-        callbackWhenSuccess = (response) => emit(loadedState.call(response.response!));
+        callbackWhenSuccess =
+            (response) => emit(loadedState.call(response.response!));
       }
     }