+++ lib/src/exceptions/src/request_canceled_exception.dart
@@ -1,0 +1,6 @@
+import 'package:spargo_tools/src/exceptions/src/app_base_exception.dart';
+
+class RequestCanceledException extends AppBaseException {
+  RequestCanceledException(
+      {required super.type, required super.message, super.stackTrace});
+}