+++ lib/src/http_client/http_client.dart
@@ -6,10 +6,12 @@ import 'dart:typed_data';
 import 'package:http/http.dart' as http;
 import 'package:http/http.dart';
 import 'package:l/l.dart';
+import 'package:spargo_tools/src/exceptions/src/request_canceled_exception.dart';
+import 'package:spargo_tools/src/http_client/cancel_completer.dart';
 import 'package:spargo_tools/src/shared_preferences_base_repository.dart';
 
-import 'app_base_settings.dart';
-import 'exceptions/exceptions.dart';
+import '../app_base_settings.dart';
+import '../exceptions/exceptions.dart';
 
 enum RequestType { get, post, put, delete }
 
@@ -25,20 +27,23 @@ abstract class AppHttp {
     String? contentType,
     bool checkToken = true,
     RequestType type = RequestType.get,
+    CancelCompleter? cancelCompleter,
   }) async {
     if (AppBaseSettings().developerDelay > 0) {
       await Future.delayed(const Duration());
     }
     // URI
     final uri = Uri.parse(
-      Uri.encodeFull((basicUrl ?? AppBaseSettings().apiUrl) + url).replaceAll('+', '%2B'),
+      Uri.encodeFull((basicUrl ?? AppBaseSettings().apiUrl) + url)
+          .replaceAll('+', '%2B'),
     );
     // Body
     String body = '';
     if (isXwwwForm && data != null) {
       final listItems = [];
       (data as Map<String, dynamic>).forEach((key, value) {
-        final valueEncoded = Uri.encodeFull(value.toString()).replaceAll('+', '%2B');
+        final valueEncoded =
+            Uri.encodeFull(value.toString()).replaceAll('+', '%2B');
         listItems.add('$key=$valueEncoded');
       });
       body = listItems.join('&');
@@ -54,27 +59,30 @@ abstract class AppHttp {
     // Headers
     final headers = <String, String>{
       'Accept': '*/*',
-      'Content-Type': isXwwwForm ? 'application/x-www-form-urlencoded' : contentType ?? ContentType.json.toString(),
+      'Content-Type': isXwwwForm
+          ? 'application/x-www-form-urlencoded'
+          : contentType ?? ContentType.json.toString(),
       'Authorization': await SharedPreferencesBaseRepository.getToken() ?? '',
       if (cookie != null) 'Cookie': cookie,
     };
     if (AppBaseSettings().logRequest) {
       log(headers.toString());
     }
+    final httpClient = http.Client();
     try {
       late Future<Response> request;
       switch (type) {
         case RequestType.get:
-          request = http.get(uri, headers: headers);
+          request = httpClient.get(uri, headers: headers);
           break;
         case RequestType.post:
-          request = http.post(uri, headers: headers, body: body);
+          request = httpClient.post(uri, headers: headers, body: body);
           break;
         case RequestType.put:
-          request = http.put(uri, headers: headers, body: body);
+          request = httpClient.put(uri, headers: headers, body: body);
           break;
         case RequestType.delete:
-          request = http.delete(uri, headers: headers, body: body);
+          request = httpClient.delete(uri, headers: headers, body: body);
           break;
       }
 
@@ -83,26 +91,50 @@ abstract class AppHttp {
         await checkUserToken();
       }
 
-      final response = await request.timeout(
-        Duration(seconds: timeout ?? AppBaseSettings().apiRequestTimeout),
-        onTimeout: () {
-          throw ServerTimeoutException(message: '', type: ExceptionType.timeout, stackTrace: StackTrace.fromString(request.toString()));
-        },
-      );
+      final response = await Future.any([
+        if (cancelCompleter != null)
+          cancelCompleter.whenCanceled.then((value) => throw (value)),
+        request.timeout(
+          Duration(seconds: timeout ?? AppBaseSettings().apiRequestTimeout),
+          onTimeout: () {
+            throw ServerTimeoutException(
+                message: '',
+                type: ExceptionType.timeout,
+                stackTrace: StackTrace.fromString(request.toString()));
+          },
+        )
+      ]);
       if (AppBaseSettings().logRequest) {
         l.i(response.body);
       }
       handleApiError(response, abortOnBadRequest);
+      httpClient.close();
       return response;
     } on SocketException catch (e, stackTrace) {
       l.e(e.message, stackTrace);
-      throw ServiceUnavailableExcecption(message: e.message, stackTrace: stackTrace, type: ExceptionType.noApi);
+      httpClient.close();
+      throw ServiceUnavailableExcecption(
+          message: e.message,
+          stackTrace: stackTrace,
+          type: ExceptionType.noApi);
     } on HandshakeException catch (e, stackTrace) {
       l.e(e.message, stackTrace);
-      throw ServiceUnavailableExcecption(message: e.message, stackTrace: stackTrace, type: ExceptionType.noApi);
+      httpClient.close();
+      throw ServiceUnavailableExcecption(
+          message: e.message,
+          stackTrace: stackTrace,
+          type: ExceptionType.noApi);
     } on ClientException catch (e, stackTrace) {
       l.e(e.message, stackTrace);
-      throw ServiceUnavailableExcecption(message: e.message, stackTrace: stackTrace, type: ExceptionType.noApi);
+      httpClient.close();
+      throw ServiceUnavailableExcecption(
+          message: e.message,
+          stackTrace: stackTrace,
+          type: ExceptionType.noApi);
+    } on RequestCanceledException catch (e, stackTrace) {
+      l.e(e.message, stackTrace);
+      httpClient.close();
+      rethrow;
     }
   }
 
@@ -115,10 +147,12 @@ abstract class AppHttp {
     String? basicUrl,
   }) async {
     final uri = Uri.parse(
-      (Uri.encodeFull((basicUrl ?? AppBaseSettings().apiUrl) + url)).replaceAll('+', '%2B'),
+      (Uri.encodeFull((basicUrl ?? AppBaseSettings().apiUrl) + url))
+          .replaceAll('+', '%2B'),
     );
     final request = http.MultipartRequest('POST', uri);
-    request.headers['Authorization'] = await SharedPreferencesBaseRepository.getToken() ?? '';
+    request.headers['Authorization'] =
+        await SharedPreferencesBaseRepository.getToken() ?? '';
     for (final key in fields.keys) {
       request.fields[key] = '${fields[key]}';
     }
@@ -136,7 +170,8 @@ abstract class AppHttp {
     required String filename,
   }) async {
     final uri = Uri.parse(
-      (Uri.encodeFull((basicUrl ?? AppBaseSettings().apiUrl) + url)).replaceAll('+', '%2B'),
+      (Uri.encodeFull((basicUrl ?? AppBaseSettings().apiUrl) + url))
+          .replaceAll('+', '%2B'),
     );
 
     final headers = <String, String>{
@@ -188,76 +223,100 @@ abstract class AppHttp {
       case HttpStatus.unauthorized:
         logErrorInfo(response);
 
-        throw UnauthorizedException(message: 'Unauthorized', type: ExceptionType.noAuth);
+        throw UnauthorizedException(
+            message: 'Unauthorized', type: ExceptionType.noAuth);
       case HttpStatus.forbidden:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Forbidden', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Forbidden', type: ExceptionType.noApi);
       case HttpStatus.notFound:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Not Found', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Not Found', type: ExceptionType.noApi);
       case HttpStatus.methodNotAllowed:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Method Not Allowed', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Method Not Allowed', type: ExceptionType.noApi);
       case HttpStatus.notAcceptable:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Not Acceptable', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Not Acceptable', type: ExceptionType.noApi);
       case HttpStatus.proxyAuthenticationRequired:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Proxy Authentication Required', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Proxy Authentication Required',
+            type: ExceptionType.noApi);
       case HttpStatus.requestTimeout:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Request Timeout', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Request Timeout', type: ExceptionType.noApi);
       case HttpStatus.conflict:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Method Not Allowed', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Method Not Allowed', type: ExceptionType.noApi);
       case HttpStatus.gone:
         logErrorInfo(response);
         throw OtherClientException(message: 'Gone', type: ExceptionType.noApi);
       case HttpStatus.lengthRequired:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Length Required', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Length Required', type: ExceptionType.noApi);
       case HttpStatus.preconditionFailed:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Precondition Failed', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Precondition Failed', type: ExceptionType.noApi);
       case HttpStatus.requestEntityTooLarge:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Request Entity Too Large', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Request Entity Too Large', type: ExceptionType.noApi);
       case HttpStatus.requestUriTooLong:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Request-URI Too Long', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Request-URI Too Long', type: ExceptionType.noApi);
       case HttpStatus.unsupportedMediaType:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Unsupported Media Type', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Unsupported Media Type', type: ExceptionType.noApi);
       case HttpStatus.requestedRangeNotSatisfiable:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Requested Range Not Satisfiable', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Requested Range Not Satisfiable',
+            type: ExceptionType.noApi);
       case HttpStatus.expectationFailed:
         logErrorInfo(response);
-        throw OtherClientException(message: 'Expectation Failed', type: ExceptionType.noApi);
+        throw OtherClientException(
+            message: 'Expectation Failed', type: ExceptionType.noApi);
 
       /// 5**
       case HttpStatus.internalServerError:
         logErrorInfo(response);
-        throw ServerException(message: 'Internal Server Error', type: ExceptionType.noApi);
+        throw ServerException(
+            message: 'Internal Server Error', type: ExceptionType.noApi);
       case HttpStatus.notImplemented:
         logErrorInfo(response);
-        throw ServerException(message: 'Not Implemented', type: ExceptionType.noApi);
+        throw ServerException(
+            message: 'Not Implemented', type: ExceptionType.noApi);
       case HttpStatus.badGateway:
         logErrorInfo(response);
-        throw ServerException(message: 'Bad Gateway', type: ExceptionType.noApi);
+        throw ServerException(
+            message: 'Bad Gateway', type: ExceptionType.noApi);
       case HttpStatus.serviceUnavailable:
         logErrorInfo(response);
-        throw ServiceUnavailableExcecption(message: 'Service Unavailable', type: ExceptionType.noApi);
+        throw ServiceUnavailableExcecption(
+            message: 'Service Unavailable', type: ExceptionType.noApi);
       case HttpStatus.gatewayTimeout:
         logErrorInfo(response);
-        throw ServerException(message: 'Gateway Timeout', type: ExceptionType.noApi);
+        throw ServerException(
+            message: 'Gateway Timeout', type: ExceptionType.noApi);
       case HttpStatus.httpVersionNotSupported:
         logErrorInfo(response);
-        throw ServerException(message: 'HTTP Version Not Supported', type: ExceptionType.noApi);
+        throw ServerException(
+            message: 'HTTP Version Not Supported', type: ExceptionType.noApi);
 
       default:
         logErrorInfo(response);
-        throw UnknownException(message: 'Unknown Exception', type: ExceptionType.noApi);
+        throw UnknownException(
+            message: 'Unknown Exception', type: ExceptionType.noApi);
     }
   }
 
@@ -388,6 +447,8 @@ abstract class ApiRequest {
     } on UnauthorizedException catch (e) {
       exception = e;
     } on ServerTimeoutException catch (e) {
+      exception = e;
+    } on RequestCanceledException catch (e) {
       exception = e;
     } catch (e, stackTrace) {
       exception = OtherClientException(