+++ CHANGELOG.md
@@ -13,3 +13,11 @@
 ## 0.0.4
 
 * Make some interfaces
+
+## 0.0.5
+
+* Created new model CancelCompeleter
+* Added test for CancelCompeleter
+* Added new exception RequestCanceledException
+* Added the ability to cancel a request using CancelCompleter in AppHttp.apiHttpRequest
+* Changed file structure for http_client